docker: skip source checksum to avoid circular dep in the release workflow
As part of the release process, a source tarball is produced,
which includes both dockerfiles and APKBUILD. Since we want to
store a valid version in the APKBUILD, but the tarball is generated
only afterward, it is not possible to calculate the checksum in
advance.

To resolve this circular dependency, the checksum is skipped.
The release workflow will ensure that the packages and
the image are created from the same tarball that is uploaded to
the Releases page.

Signed-off-by: László Várady <laszlo.varady@anno.io>