cfg-lexer: capture original location even if @line is present
If @line is encountered in the input stream, we overwrite our idea of where
we are in the input file so that we can properly report on file locations,
even if we are expanding a block or other generators.

This patch makes it possible to recover the original location (e.g. before
@line was called), which is needed to extract the source text _after_
generation/substitution has happened.

This means that whenever we display the source text (e.g. error report,
or as a part of the filterx expr being executed), we can use
the post-processed version.

Whenever we want to read the file that contained the source, we can use
the current yylloc. If we want to extract from the postprocessed
buffer, we need to get back to the original yylloc.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>