cfg-grammar: fix the precedence of ';'
In 6d74e60c4b2b51e9842ccab23902803b9ff8de2b, the precedence of ';' has
implicitly changed for filter rules, changing the meaning of filter
expressions:

The expression `not program("a"); message("b");` should have the following
interpretation: `(not program("a")) and message("b")`.

Due to accidentally changed precedence of ';', currently it is interpreted
as `not (message("a") and message("b"))`.

This patch recovers the original precedence, but also introduces
precedence and left associativity for other semicolon-separated rules.

All the other (non-filter) uses of the ';' token are unambiguous, so this
declaration won't interfere with those rules.

Signed-off-by: László Várady <laszlo.varady@anno.io>
Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>