wildcard-file: fix race between EOF and file deletion detection
This patch fixes a "race condition" between the EOF and file deletion
detection of AxoSyslog when wildcard-file() is used.

If a file is written after being deleted (an application keeps an fd
open), or if these two events occur concurrently, nearly at the same
time, the file poller mechanism might schedule another read cycle while
the file has already been marked as fully read and deleted.

To avoid scheduling issues between the two checks, the following is done
from now on: We do not track the EOF state in the WildcardFileReader,
instead, when a file deletion notification is received, the poller will
be instructed to stop after reaching the next EOF, and only then will
AxoSyslog actually delete the reader instance.

Signed-off-by: László Várady <laszlo.varady@anno.io>