filterx: add metrics-labels class
This class is responsible for formatting a `FilterXExpr`
to an array of `StatsClusterLabel`s.

The expression must be evaluate to a `Dict[str, Any]`.
The formatting is cached wherever it is possible.

The implementation uses TLS metrics cache, as `FilterXExpr`s
must be stateless currently, but if the need arises, the code
can be easily changed to use an arbitrary metrics cache.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>