filterx: add metrics class
This class is responsible for managing `StatsCounterItem`s
based on `key` and `labels` `FilterXExpr`s.

`key` must evaluate to a `str`.
`labels` must evaluate to a `Dict[str, Any]` is handled by
`FilterXMetricsLabels`.

The implementation uses TLS metrics cache, as `FilterXExpr`s
must be stateless currently, but if the need arises, the code
can be easily changed to use an arbitrary metrics cache

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>