filterx: introduce expr-plus-generator operator
Introduced expr-plus-generator as an alternative to the expr-plus operator,
designed to work with generators. This new operator functions as a wrapper
around the LHS and RHS generators, which can be either generators or variables.
At least one of them needs to be a generator. Both generator and variable
elements will be handled as a single generator in further processing.

Signed-off-by: shifter <shifter@axoflow.com>