filterx,cfg-lexer: pre-validate function names in lexer
We will introduce generator functions in the next commit.
Generator functions have the exact same syntax as regular
functions, but we need to write separate rules in the grammar
to handle them, but if we don't know whether an identifier
is a regular function name or generator function name, the
rules will conflict.

We can solve this by preprocessing LL_IDENTIFIERs in the
lexer, and trying to find a corresponding function for it,
and if it exists, returning a separate token: LL_FILTERX_FUNC.

Some of this commit's code changes on their own might not really
make sense, but it will in the next commit.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>