msg-format: add piggyback-errors and no-piggyback-errors flags
This tells syslog-ng not to wipe out the input message in syslog-parser() and
source encapsulated syslog parsers when there's a parse issue.

For example:

    parser { syslog-parser(flags(syslog-protocol, no-piggyback-errors)); };

With that in place, $MSG retains its original value and various tags (e.g.
the syslog.* ones) report what kind of issues we detected with the messsage.

When the message is sent to a syslog-like destination, we will prepend
a header.

Basically this makes 5424 and 3164 parsing more similar. Previously 5424 parsing
reported its errors by completely wiping out the the existing log message
fields and adding a message attributed to "syslog-ng" itself. But this
message was coming from the same source and not internal().




Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>