syslog-format: get rid off RFC3164 error case in favour of setting error tags
The only case we had a hard error in RFC3164 was when the incoming
data had an opening angle bracket ("<") without a closing one.

This almost never happens, so the filtering out errors case is not a strong
one, and like with any other errors, it is better to accept these messages.
In this case, the entire input line is put into $MSG and a new error
tag is set (syslog.invalid_pri).

The old behaviour of setting $MSG to "Error processing message: ..." that
is coming from syslog-ng itself was not very intuitive anyway and that
did not happen for real world errors anyway.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>