filterx: rework of unset_empties
- Default empty set: "", null, [], {}.
- Optional named arguments:
  - recursive: Enables recursive processing of nested dictionaries.
  - ignorecase: Enables case-insensitive matching.
  - replacement: Specifies a filterxobject to replace target elements instead of removing them.
  - targets: A list of elements to identify for removal or replacement, clearing the default empty set.
- UTF-8 support.
- ignorecase and recursive args are defaults to TRUE

This change removes elements from the given dictionary or list that match
the empty set. If the recursive argument is provided, the function will
process nested dictionaries as well. The replacement argument allows
replacing target elements with a specified object, and the targets
argument customizes which elements are removed or replaced, overriding
the default empty set.

Signed-off-by: shifter <shifter@axoflow.com>