filterx: support passing generator to functions' not first param
Unfortunately, the syntax of function call with a generator
as its first param is reserved for the generator casted
assignment. Sometime we should fix conflict fully, but for
now supporting generators starting from the second param
is a bit better than not supporting them at all.

(Will be useful for startswith(), endswith())

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>