cfg: fix config-version incosistencies
Template functions, for example, are constructed right when AxoSyslog
parses the configuration, so the `@version` directive must be placed at
the top of the config file to make it consistent.

This patch fixes the implicit "current version" mechanism (when no
`@version` is specified) by moving it back to the lexer, and warns the
user about misplaced or redefined `@version` directives.

Signed-off-by: László Várady <laszlo.varady@anno.io>