filterx: add FilterXRef
Deep-copy assignment (the `.something = `, `["something"] = `,
`something =`, `something +=` operators) is an essential feature of the
FilterX language, making it easy to trace and understand how values
change.

However, copy (especially deep-copy) is expensive, so this PR aims to
fix the performance bottleneck by introducing copy-on-write for mutable
objects.

To achieve CoW semantics without introducing endless changes in the
`filterx` folder, `FilterXRef` has been introduced. References are
currently not part of the FilterX language (hopefully, they never will
be). `FilterXRef` is used only internally to reference the same
`FilterXObject` from multiple locations (variables, other types of
assignments). For this reason, `FilterXRef` pretends to be a
`FilterXObject`, but it's not a real FilterX type.

`FilterXRef` transparently propagates operations (virtual function
calls) to the referenced value, but it also makes sure that the value is
copied on the first write operation in case multiple references exist.

`FilterXRef` is created only for mutable objects, and only when they are
needed: around assignment-like operators; this and its `FilterXObject`
behavior was the key to avoid polluting all code paths with references.

Signed-off-by: László Várady <laszlo.varady@anno.io>