filterx: create references where objects are assigned to names
`FilterXRef` is created only for mutable objects, and only when they are
needed: around assignment-like operators; this and its `FilterXObject`
behavior is the key to avoid polluting all code paths with references.

Full CoW on subscript and attr operators (`.something`, `["something"]`)
of JSON objects is currently not implemented. I've tried to make it
work, but it requires a complete redesign of how we cache JSON
subobjects (or alternatively, we need an internal representation for
dicts and lists without using the json-c library). This will be
implemented later. For now, such operators make unnecessary copies (CoW
only partially works on them).

Signed-off-by: László Várady <laszlo.varady@anno.io>