filterx: unwrap references where downcasts are used
FilterXRef can mimic the functionality of its referenced value, but only
through virtual methods.

Where (implicit) downcasts are used, it is impossible to be transparent
about references, so these unwrap methods have to be used.
This is far from ideal, but doing downcasts based on is_type() is the
original and real problem here.

In such cases, `filterx_ref_unwrap()` must be called to extract the
original object. The extracted object must only be used locally
(short-term, without referencing or storing it anywhere).

Signed-off-by: László Várady <laszlo.varady@anno.io>