lib/filterx: fix variable lookup
A non-floating (i.e., tied to a LogMessage) FilterX variable is considered invalid if:
- its' generation counter is 0 AND
- the scope is syncable

This is to allow two consecutive FilterX blocks to pass non-floating variables
without calling filterx_scope_sync().
If they would be considered invalid then the changes would be lost in the latter block.
This is because it would fetch it via _pull_variable_from_message() which would
yield the unmodified LogMessageValue.

Signed-off-by: Szilard Parrag <szilard.parrag@axoflow.com>