filterx: support literal gen API on literal inner gen
The inner generator part of the code is a workaround
for the grammar, and is not very clean to have.

We would like to iterate through it in some expr
ctors, like it was a normal literal generator.

Also the len() function could be useful somewhere.

Removing the literal inner generators would be a
bigger task, so I just hacked the literal generator
API a bit to support literal inner generators.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>