filterx: implement set_fields()
If there are overrides defined for the field,
we first iterate through them, and if there is
an expression that successfully evaluated, we
set it.

If overrides were not effective, and the field
already has a value, we are done with this field,
we leave it as it is.

If, however it did not have a value, we iterate through
the defaults, like we did with the overrides, and set
the field accordingly.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>