light: fix bytes/protobuf testcases
With the $MSG as a macro gone, $MSG became a typed value, so it does matter
what we set it to. Since setting it to a bytes/protobuf value means
that its not a string anymore, we can't render it as a part of a template
unless we explicitly ask for the right type hint.

This is what this patch changes in the broken testcases.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>