filterx/filterx-scope: retain the generation counter across clone operations
Previously scope->generation was reset to 0 in case we ended up doing a
scope clone operation.

Let's retain the generation counter instead, so we don't need to adjust the
generation value for FilterXValues either.  Also, subsequent patches will
start sharing FilterXVariable instances, meaning we want to know which
generation a specific variable is associated with.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>