cfg-source: cache the result of g_strsplit()
When extracting the source text, we need to split the input into lines.
Previously this was done every time we extracted a line of text, causing
the startup to be very slow when I enabled source extraction for
every FilterXExpr instance.

This patch causes the result of that g_strsplit() call to be saved,
effectively using more memory startup, but being a lot more efficient (e.g.
startup within a second instead of 20 seconds or so).


Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>