filterx: don't use FilterXString in interface
The convention is that we are using FilterXObject pointers in interfaces,
and potentially check if the object at runtime matches our expectations.
By using FilterXString instances there, we typed constructors as well
as publish FilterXString as a struct. I am not against publishing
internal structs if they serve optimization purposes, but any value should
rather be passed around as objects and not specific types.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>