generic-number: make the gint64 path in the functions inline
We are encapsulating a GenericNumber instance in our integer/boolean/double
objects, basically having two layers of encapsulation.

I attempted to remove one of these but it was a lot more difficult and would
require API changes, so for now I was satisfied with inlining the
integer paths, which is exercised whenever we are using an integer
in filterx (e.g. iterating over a list in parse_csv()).

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>