filterx/json-repr: add json-repr module to implement JSON based serialization of filterx objects
This change is the start making JSON just an external representation of
FilterXObjects instead of a native storage mechanism.

This module incorporates both parsing JSON into FilterXObjects and
serializing FilterXObjects into JSON strings.

It exposes both a C and filterx functions to perform these operations, which
also means that format_json() is pulled into the core instead of being part
of the json plugin.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>