filterx/expr-function: simple functions should get unmarshalled values
Right now, arguments to simple functions are evaluated at the time of the
call, but that means that FilterXMessageValue objects are encountered when
evaluating them.

This is:
1) inconvinient, as we have to consider both native objects and
   message values
2) somewhat incorrect, as the unmarshalled version of the arguments are
   not cached in a filterx variable.
3) if the function changes the copy, it will not be impacting the
   filterx variable, e.g. unset_empties($json) will be a noop, if $json
   is an LM_VT_JSON

I understand this may have a performance impact, but if it does, we can
look into creating FilterXStrings with a borrowed string pointer.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>