filterx: native dict/list types
This patch reworks the data storage behind FilterXDictObject and adds
FilterXListObject implementation to be the "native" storage backends for
filterx.

With that the other dict/list implementations (notably json) will be less
"native" and might be eliminated completely, along with generators and other
complexities that allowed the selective use of any data backends.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>