filterx/filterx-ref: allow setting "readonly" through a ref
All mutable objects are to be tracked through refs (in subsequent patches),
which means that cached_json_file() will also need to work with ref-values,
while it freezes them.

Instead of prohibiting the use of filterx_object_make_readonly() in a
FilterXRef instance, forward these calls to the underlying value.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>