filterx/filterx-expr: propagate the need for a writable object
In the case of setattr/set_subscript, we need an object that is writable,
e.g. one where copy-on-write already happened.

This patch implements an infrastructure, where an expr can request a child
expr to return a writable instance, so that we propagate this back to the
root of the expr tree.

With that information, expr-variable can trigger the copy before returning
the variable.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>