appmodel/transformation: add parser support (confgen)
This is a rather big change, and unfortunately
I could not split it into smaller parts.

---

The logic is the following:

Transformations can have either filterx only transforms
or mixed/parser only transform.

First, we gather the transformations that have mixed/parser
only transforms, and first check them with log level if-elifs
and a filterx pre-checks to the variable and app name.

Then, to keep the filterx switch-case optimization,
we gather filterx transformations that have filterx
transforms active only (include/exclude-transforms() honored)
and put them to an else block to a switch-case.

There are cases where no mixed/parser only or no filterx only
transformations are available, those are handled correctly, too.

---

Heavy E2E tests will be added in the next commit.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>