filterx/filterx-ref: all mutable objects should be wrapped by refs
All of
  * variables,
  * dictionary values
  * list elements

should always be wrapped into a ref, as dirty tracking for mutable values is
based on refs. JSON is an exception, as child objects embedded in other objects
are not using references, only the root element is refd. But the plan is to
eliminate json-c based storage completely.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>