filterx/filterx-object: allow reference counted freeze/unfreeze operations
Sometimes the same object is freezed multiple times.  For instance, a cached
string instance (e.g.  the zero length string) is frozen by the interpreter
at startup.  Then the same string instance can be inside the dict frozen by
cached_json_file().

In this case we will have two freeze operations on the same object, thereby
two unfreeze operations as well.

There would be two alternatives to track this:
  1) the caller remembers if the freeze operation was succesful and only
     unfreeze in the success case.

  2) the object can track the number of freeze operations and pair them up
     with unfreezes.

The first would require an extra bit on the user of the frozen object, which
is sometimes difficult to store, so I chose the 2nd alternative.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>