light: store testrunuid in session data
Sometimes with unexpected process stops the
SessionData file can stay on the file system with
initialized values in it.

This commit uses the PYTEST_XDIST_TESTRUNUID env
var, which is provided by xdist for workers, and
is guaranteed to be unique for each run, and be
the same for all workers for that run.

We store this value to the SessionData file, and
during start if we find a file that has a
different uid, we clear its contents, and start a
new one.

With this we also restart the active worker count,
and will be able to delete the file after the run.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>