filterx/filterx-ref: do not provide an unmarshal function
Refs only wrap mutable objects, unmarshal is only there to deserialize
FilterXMessageValue objects, which is not mutable in place.

This means that this unmarshal function never really performs a function,
however it still is invoked and then returns any time a ref-wrapped
object is referenced. Removing the unmarshal method gave me 200msec savings
on a 3.7s baseline in my tests, e.g. ~5%.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>