filterx/filterx-weakref: add filterx_weakref_invoke()
Instead of having to take a ref to the weakrefd object, call
filterx_weakref_invoke() which can invoke a function without having to do
that.  filterx_object_ref/unref was actually the most expensive call in the
copy-on-write implementation for dicts, which iterates over the dict
hierarchy to clone dicts as needed.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>