filterx/object-dict: fix unset in case of a hash collision
In case an item was unset() while another item with the same hash is
in the dict, we erroneously reported the other item as missing as well.

The problem is that we can only break out of the lookup loop in case the
index is its initial value (FXD_IX_EMPTY) and not when it is cleared
(e.g. FXD_IX_DUMMY). In the latter case, we need to continue the lookup
loop.

However, we should try to reuse the cleared slot for the new item.
We do this by storing the first found dummy slot and returning it
if we did not find any valid values in the lookup loop.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>
Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>