filterx/filterx-scope: write-protected FilterXScope
By making FilterXScope initially write protected and only making it writable
at the first actual write, we can delay cloning of variables for read-only
access.

For instance, given foo is a dict:

foo = {"foo":{"foo2":"foovalue"}};

And then in a subsequent filterx block we do this:

foo.foo.foo2

Then the following happens without this patch:
  * "foo" is fetched from the parent scope
  * filterx_scope_clone() is invoked, which will create another ref
  * we continue evaluation

This means that unmatched branches will cause clones. This branch changes
this by:
  * making the scope initially write protected
  * making it writable at strategic points
  * as long as it is read-only, mutable variables are shared between the
    current and the previous FilterXScope, no cloning is performed

We mark the scope as writable at the following points:
  * set-subscript, unset, assign, set-attr, update_repr

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>