transport-tls: fix cond handling
When the transport layer declares its IO needs, the proto layer considers
it vital and non-overridable.

Setting default values is therefore not correct, and more importantly:
forcing G_IO_IN in tls_read() is incorrect, because this would override
a proto's read+write need forever (for example, a proto that allows
continuously possible async reads during writes).

OpenSSL requires that operations be retried in case of a WANT_* result,
and this requirement will continue to be fulfilled with this patch.

Signed-off-by: László Várady <laszlo.varady@anno.io>