docker: only build required build-deps on arm/v7
Our arm/v7 compilation in github is slow and glib fails to build from
source at the moment. Address both of these by simply not compiling
anything but json-c.

json-c is required as we need to enable threading support. The list
of build deps to build is embedded in the dockerfile for local build
support but is overridden from the github workflow.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>