filterx: print every error in stack
I chose to print the errors starting from the ones
deeper in the actual call chain. For example we
have a binary expression which evaluates lhs.
If lhs fails, the lhs eval error will be printed
sooner, than the binary expression's.

I also print an error index ascending from 1 to
the number of errors (e.g. [1/2], [2/2]).

I think these are sensible, as we have a lot more
other logs around these, where the events
happening sooner are logged sooner, so I tried to
mimic that, but we can choose a different printing
if needed.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>