cef: add support for changing the header fields during parsing or formatting
We have parsing/unescaping bugs, which need
bigger modifications to the code.

The current `optional` flag for fields makes the
code too complex to introduce changes to.

The following commits will make an alternative
solution: separate LEEF v1 and v2 fields, and
switching between them parse or format time.

This commit starts to prepares this.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>