filterx/object-metrics-labels: do not store null values in labels
We won't ever render them in the final metric's
labels, so no need to store them.

This is not just a perf optimization, but also a
bugfix, as storing null values in labels resulted
in the value "null", which we did not drop in
update_metrics(), like we wanted to.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>