mainloop-io-worker: use spearate work pools for source/destination/processing
Prior to this patch we were only using a single workpool for all kinds
of I/O events (source, destination and processing). However this can
generate extra thread migrations due to different tasks scheduled to the
same thread.

For example, if a source thread is waking up a destination, it might
end up taking the same thread that the source was using. Meaning, the
source will be scheduled to a different thread, losing its cache
contents, etc.

With a separate work pool, the destination task will never take the
same thread as the source task, so thread affinity should be improved.

With this change, we will oversubscribe the number of CPUs, as each work
pool will take the same max_threads value. This will ultimately be
resolved by the kernel scheduler. Scaling would suffer once we actually
execute in more threads than available CPUs, but normally threads are only
allocated on-demand.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>