cfg-lexer: fix slash_string escaping
Let's redefine what slash_string means in order to make the escaping rules
clear:

Everything between two / characters are treated literally, except for \/,
which can be used to include a / character within the string.
If you want to include \/ within the string, you can do that too: \\/

This type of string literal syntax can be used to construct regexp
patterns.

pcre2 handles special non-printing characters (\r, \n, \t, etc.),
octals (\oDDD), and hex as well (\xDD), so no need to escape those in our
lexer.

Note: it is not possible to represent strings that end with a single \
but that would be an invalid regexp, so this limitation is acceptable.

Signed-off-by: László Várady <laszlo.varady@axoflow.com>