configure.ac: add JSON_LIBS to mixed mode dependencies as well
"mixed" mode is probably unused by now, it was used to build syslog-ng
so that it didn't use anything outside of / (e.g. /usr/lib/libssl.so),
but distros have merged / and /usr fortunately, so this is not an
issue anymore.

Anyway, JSON_LIBS was missing from this path, so add it back.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>