filterx/filterx-scope: don't pull variables from parent if read-only
By making FilterXScope initially write protected and only making it writable
at the first actual write, we can delay cloning of variables for read-only
access.

For instance, given foo is a dict:

    foo = {"foo":{"foo2":"foovalue"}};

And then in a subsequent filterx block we do this:

    foo.foo.foo2

Then the following happens without this patch:
  * "foo" is fetched from the parent scope
  * filterx_scope_clone() is invoked, which will create another ref
  * the ref is added to the scope
  * we continue evaluation

This means that initial filterx expressions that drop the majority of the
data will create this ref even if the condition does not match.  This is not
a real problem if all or almost all of the data is accepted by these
conditions, as in those cases we would probably need those clones anyway,
but present an overhead if:
  * most of the data is dropped
  * we have many of these unmatching branches.

With the read-only/read-write tracking of the scope, we change the behaviour
as follows:
  * we make the scope initially write protected
  * we make it writable on-demand (e.g. when we actually change a variable in the scope)
  * as long as it is read-only, refs are not created, even otherwise mutable
    variables are shared between the current and the previous
    FilterXScope, no cloning is performed

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>