filterx/object-{string,primitive}: add FILTERX_*_CLEAR_FROM_STACK()
Instead of merely calling filterx_object_unref() on the stack allocated
instance (which is a noop), check if it has changed first.

This conditional can be proven to be always false or always true,
eliminating the entire unref call for variables that are never changed,
which is pretty usual for indexes and local values.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>