filterx/expr-variable: fix assignment tracking for unmarshalled variables
If a just assigned variable is unmarshalled in place, the "assigned"
value was always set to FALSE, even if it was originally TRUE. This means
that these variables were to synced to the message at the end of
the filterx block.

Here's an example:

                $foo = $PROGRAM;
                if ($foo === "whatever") {
                }

The condition unmarshalls the value, causing the "assigned" value for $foo
to be set to FALSE.

This means that filterx_scope_sync() will not write this value to
the message.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>