logthrdestdrv: add worker-partition-buckets() option
This patch adds a new worker-partition-buckets() option along with the
existing worker-partition-key(). This takes a template that tells
the threaded fanout algorithm how many different buckets we should use
for the same partition key.

E.g. if the partition-key is $PROGRAM and a lot of data is generated by
the same $PROGRAM, we may want to spread that to multiple threads, instead
of just one.

The template specified to worker-partition-buckets() should expand to an
integer that specifies how many different buckets we want to use.

The current algorithm in LogThreadedDestDriver is to use the entropy
in the receive timestamp to spread out the messages to different workers.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>