filterx: fix potential use-after-free
The payload member is sometimes piggybacked to the end of the LogMessage
struct, so even if we are taking a reference to it (using nv_table_ref)
to prevent it being freed, it still is freed, if the underlying
LogMessage itself is destroyed.

This patch introduces a new API, log_msg_pin_payload() and log_msg_unpin_payload()
that takes the required references thereby ensuring that the LogMessage
is not freed either, at least as long as we have active pins.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>