dyn-metrics-store: fix race on StatsCounter use_count
`use_count` is a non-atomic integer.

Modifying it from parallel threads at the same time
is racy.

Both `merge()` and `retrieve_counter()` increases
the `use_count` of a counter.

- `merge()` is called with the `global_metrics_cache_lock` held
- `retrieve_counter()` calls `stats_lock()` inside

If `merge()` is being called in one thread (thread deinit)
and `retrieve_counter()` is being called in another
(new counter for the TLS cache), they can run in the same
time, incrementing `use_count` in a racy manner.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>