filterx/filterx-eval: truncate weakref array at the end of a filterx block
The weakref array can only contain objects that were created by the current
or subsequent filterx blocks.

This means as we are exiting, we can free everything that is not needed
anymore.  Apart from the memory use, these weakrefs can also hurt
performance: they may keep extra FilterXRefs around that point to mutable
objects.  When changing those, that will trigger a copy-on-write.  By
freeing the unused references, we can avoid the copy.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>