logmsg: free message ref in log_msg_free_queue_node()
To make log_msg_alloc_queue_node() and log_msg_free_queue_node() consistent
reference counting wise, we need to the ref in alloc_queue_node(), which is the
case even now and drop the same ref in log_msg_free_queue_node(), which
was not the case.

This patch adds the missing log_msg_unref() and adapts all the code to
use this assumption.

This also fixes a potential use-after-free in log_queue_fifo_drop_messages_from_input_queue()
as we were accessing node->msg _after_ freeing the queue node.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>